+++ src/main/application.ts
@@ -1,12 +1,11 @@
 import { Container } from 'inversify';
-import { RequestFactory } from './request';
 import { Logger, StandardLogger } from './logger';
-import { Configuration, EnvConfiguration } from './config';
 import { HttpServer } from './http';
 import { MetricsRouter } from './metrics/route';
 import { Router } from './router';
 import { MetricsRegistry } from './metrics';
 import { getGlobalMetrics } from './metrics/global';
+import { FrameworkEnv } from './env';
 
 /**
  * Application provides an IoC container where all modules should be registered
@@ -31,18 +30,13 @@ export class Application {
         this.container.bind('RootContainer').toConstantValue(container);
         this.container.bind(HttpServer).toSelf().inSingletonScope();
         this.container.bind(Logger).to(StandardLogger).inSingletonScope();
-        this.container.bind(Configuration).to(EnvConfiguration).inSingletonScope();
-        this.container.bind(RequestFactory).toSelf();
         this.container.bind(Router).to(MetricsRouter);
         this.container.bind(MetricsRegistry).toConstantValue(getGlobalMetrics());
+        this.container.bind(FrameworkEnv).toSelf().inSingletonScope();
     }
 
     get logger(): Logger {
         return this.container.get(Logger);
-    }
-
-    get config(): Configuration {
-        return this.container.get(Configuration);
     }
 
     get httpServer(): HttpServer {