+++ src/main/jwks.ts
@@ -1,0 +1,112 @@
+import Ajv from 'ajv';
+import { Request, Fetch } from '@automationcloud/request';
+import { Exception } from './exception';
+import { ajvErrorToMessage } from './util';
+
+const ajv = new Ajv({
+    allErrors: true,
+    useDefaults: true,
+    jsonPointers: true,
+    format: 'full',
+});
+
+// eslint-disable-next-line import/no-commonjs
+const validateFunction = ajv.compile(require('../../schema/ac-jwks.json'));
+
+export class JwksClient {
+    protected _cache: JwksCache | null = null;
+    request: Request;
+
+    constructor(protected options: JwksOptions) {
+        this.options = options;
+        this.request = new Request({
+            baseUrl: this.options.url,
+            retryAttempts: this.options.retryAttempts ?? 3,
+            fetch: this.options.fetch,
+        });
+    }
+
+    async getSigningKey(): Promise<string> {
+        const keys = await this.getSigningKeys();
+        const alg = this.options.algorithm;
+        const matchingKey = keys.find(k => k.alg === alg);
+        if (!matchingKey) {
+            throw new Exception({
+                name: 'SigningKeyNotFoundError',
+                message: 'Expected signing key not found from response',
+            });
+        }
+
+        return matchingKey.k;
+    }
+
+    async getSigningKeys(): Promise<SigningKey[]> {
+        const { keys, validUntil = 0 } = this._cache || {};
+        if (keys && Date.now() < validUntil) {
+            return keys;
+        }
+
+        const res = await this.request.get('');
+        const validRes = this.validateResponse(res);
+        this.setCache(validRes.keys);
+
+        return validRes.keys;
+    }
+
+    getCache() {
+        return this._cache == null ? null : { ...this._cache };
+    }
+
+    setCache(keys: SigningKey[], maxAge?: number) {
+        const ttl = maxAge ??
+            this.options.cacheMaxAge ??
+            60 * 1000;
+
+        const validUntil = Date.now() + ttl;
+        this._cache = { keys, validUntil };
+    }
+
+    clearCache() {
+        this._cache = null;
+    }
+
+    protected validateResponse(res: { [k: string]: any }): SigningKeySets {
+        if (validateFunction(res) === true) {
+            return res as SigningKeySets;
+        }
+
+        const errors = validateFunction.errors || [];
+        const messages = errors.map(e => ajvErrorToMessage(e));
+
+        throw new Exception({
+            name: 'JwksValidationError',
+            message: 'Automation cloud jwks response validation failed',
+            details: {
+                messages,
+            }
+        });
+    }
+}
+
+export interface JwksOptions {
+    url: string;
+    algorithm: string;
+    cacheMaxAge?: number;
+
+    retryAttempts?: number;
+    fetch?: Fetch;
+}
+
+export type JwksCache = SigningKeySets & { validUntil: number }
+
+export interface SigningKeySets {
+    keys: SigningKey[]
+}
+
+export interface SigningKey {
+    use?: string;
+    kty?: string;
+    kid?: string;
+    alg: string;
+    k: string;
+}