+++ src/test/specs/services/auth.test.ts
@@ -1,46 +1,80 @@
 import { Container } from 'inversify';
+import { RequestOptions, Response } from '@automationcloud/request';
+import assert from 'assert';
 import {
-    Configuration,
-    ForwardRequestHeaderAuthService,
-    RequestFactory,
     AuthService,
     Logger,
     ConsoleLogger,
-    RequestOptions,
-    JwtAuthService,
+    AutomationCloudAuthService,
+    Jwt,
+    AutomationCloudJwt,
 } from '../../../main';
-import assert from 'assert';
-import { Jwt, KeycloakJwtMock } from '../../../main/jwt';
+import { FrameworkEnv } from '../../../main/env';
 
-describe('ForwardRequestHeaderAuthService', () => {
+describe('AutomationCloudAuthService', () => {
     let container: Container;
     let requestSent = false;
     let requestHeaders: any = {};
+    let authService: AutomationCloudAuthService;
 
     beforeEach(() => {
         requestSent = false;
         requestHeaders = {};
-        const request: any = {
-            get(url: string, options: RequestOptions) {
-                requestSent = true;
-                requestHeaders = options.headers;
-            }
-        };
+        const fetchMock = (url: string, options: RequestOptions) => {
+            requestSent = true;
+            requestHeaders = options.headers;
+            return Promise.resolve(new Response('{}'));
+        }
+
         container = new Container({ skipBaseClassChecks: true });
-        container.bind(Configuration).toSelf();
         container.bind(Logger).to(ConsoleLogger);
-        container.bind(AuthService).to(ForwardRequestHeaderAuthService);
-        container.bind(RequestFactory).to(class extends RequestFactory {
-            create() {
-                return request;
-            }
-        });
+        container.bind(AuthService).to(AutomationCloudAuthService);
+        container.bind(Jwt).to(AutomationCloudJwt).inSingletonScope();
+        container.bind(FrameworkEnv).toSelf().inSingletonScope();
+        authService = container.get(AuthService) as AutomationCloudAuthService;
+        authService.request.config.fetch = fetchMock;
     });
 
-    context('authorization header valid', () => {
+    context('ctx.headers[AC_AUTH_HEADER] exists', () => {
+        beforeEach(async () => {
+            // mock decodeAndVerify
+            const jwt = container.get(Jwt) as AutomationCloudJwt;
+            jwt.decodeAndVerify = async (_token: string) => {
+                return {
+                    context: {
+                        user_id: 'some-user',
+                        organisation_id: 'some-user-org-id',
+                    },
+                    authorization: {},
+                    authentication: {
+                        mechanism: 'client_credentials'
+                    },
+                }
+            };
+
+            const authHeader = container.get(FrameworkEnv).AC_AUTH_HEADER_NAME;
+            const headers = {} as any;
+            headers[authHeader] = 'Bearer jwt-token-here';
+            const ctx: any = { req: { headers } };
+
+            authService = container.get(AuthService) as AutomationCloudAuthService;
+            await authService.authorize(ctx);
+        })
+
+        it('gets organisation_id from token', async () => {
+            const organisationId = authService.getOrganisationId();
+            assert.equal(organisationId, 'some-user-org-id');
+        });
+
+        it('does not send request to s-api', async () => {
+            assert.equal(requestSent, false);
+        });
+
+    });
+
+    context('authorization header exist & valid', () => {
 
         it('sends a request with Authorization header', async () => {
-            const authService = container.get(AuthService);
             const ctx: any = { req: { headers: { authorization: 'AUTH' } } };
             assert.equal(requestSent, false);
             await authService.authorize(ctx);
@@ -51,7 +85,7 @@ describe('ForwardRequestHeaderAuthService', () => {
         it('does not send request if Authorization is cached', async () => {
             const ttl = 60000;
             const margin = 1000;
-            const authService = container.get(AuthService) as ForwardRequestHeaderAuthService;
+
             authService.cacheTtl = ttl;
             authService.authorizedCache.set('AUTH', Date.now() - ttl + margin);
             const ctx: any = { req: { headers: { authorization: 'AUTH' } } };
@@ -63,7 +97,6 @@ describe('ForwardRequestHeaderAuthService', () => {
         it('still sends request if cache expires', async () => {
             const ttl = 60000;
             const margin = 1000;
-            const authService = container.get(AuthService) as ForwardRequestHeaderAuthService;
             authService.cacheTtl = ttl;
             authService.authorizedCache.set('AUTH', Date.now() - ttl - margin);
             const ctx: any = { req: { headers: { authorization: 'AUTH' } } };
@@ -79,68 +112,12 @@ describe('ForwardRequestHeaderAuthService', () => {
         it('throws AuthenticationError', async () => {
             const authService = container.get(AuthService);
             const ctx: any = { req: { headers: {} } };
-            assert.equal(requestSent, false);
             try {
                 await authService.authorize(ctx);
                 throw new Error('Unexpected success');
             } catch (err) {
                 assert.equal(err.code, 'AuthenticationError');
                 assert.equal(requestSent, false);
-            }
-        });
-
-    });
-
-});
-
-describe('JwtAuthService', () => {
-    let container: Container;
-    let jwtMock: KeycloakJwtMock;
-
-    beforeEach(() => {
-        container = new Container({ skipBaseClassChecks: true });
-        container.bind(Configuration).toSelf();
-        container.bind(Logger).to(ConsoleLogger);
-        container.bind(KeycloakJwtMock).toSelf().inSingletonScope();
-        container.bind(Jwt).toService(KeycloakJwtMock);
-        container.bind(AuthService).to(JwtAuthService).inSingletonScope();
-        jwtMock = container.get(KeycloakJwtMock);
-    });
-
-    it('gets actorModel and actorId from token', async () => {
-        const authService = container.get(AuthService);
-        const token = jwtMock.createToken({ userId: 'some-user', organisationId: 'some-user-org-id' });
-        const ctx: any = { req: { headers: { authorization: 'Bearer ' + token } } };
-        await authService.authorize(ctx);
-
-        assert.equal(ctx.state.actorModel, 'User');
-        assert.equal(ctx.state.actorId, 'some-user');
-        assert.equal(ctx.state.organisationId, 'some-user-org-id');
-    });
-
-    it('throws error when unexpected actor is decoded', async () => {
-        const authService = container.get(AuthService);
-        const token = jwtMock.createToken({ randomActor: 'some-user' });
-
-        const ctx: any = { req: { headers: { authorization: 'Bearer ' + token } } };
-        try {
-            await authService.authorize(ctx);
-            throw new Error('Unexpected success');
-        } catch (err) {
-            assert.equal(err.code, 'AuthenticationError');
-        }
-    });
-
-    context('authorization header does not exist', () => {
-
-        it('throws AuthenticationError', async () => {
-            const authService = container.get(AuthService);
-            const ctx: any = { req: { headers: {} } };
-            try {
-                await authService.authorize(ctx);
-                throw new Error('Unexpected success');
-            } catch (err) {
-                assert.equal(err.code, 'AuthenticationError');
             }
         });